Refactor tests
When the coach ran through the challenge, tests were failing
unexpectedly.  The coach was confused why a test failed when
it was sending the right response back.  But, because the
coach's method was destructive, it mutated the object passed
to the rspec matcher.  This eliminates that problem, so that
the first two tests are not dependent on the method being
non-destructive.